+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - make render process cancellable
 - add `userInput` error code to `MDRenderError`
 - add `abortSignal` property to `MDRenderConfig`
+- add `allowExternalFiles` property to `MDRenderConfig`
 ### Changed
 ### Deprecated
 ### Removed